# swagger.yaml
swagger: "2.0"
info:
  description: "This is the specification for the API of the Frontend-Service of the MemeIt-Project for maue-wise2021. The frontend and this API is written by Talha Dagli, John Einicke and Yoel Berendsohn."
  version: "2.1.0"
  title: "MemeIt-Frontend-Service"
host: "host.frontend.io"
basePath: "/"
...
schemes:
- "https"
paths:
  /search:
    post:
      tags:
      - "Search"
      summary: "Gets search results"
      description: "Takes a query and gives a list of search results."
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "query"
        description: "Query-String"
        required: true
        schema:
          properties:
            query:
              type: string
          example:
            query: "Junior developer memes"
      responses:
        200:
          description: "Successful operation"
          schema:
            $ref: "#/definitions/query_response"
        500:
          description: "Service error"
          schema:
            properties:
              message:
                type: string
              error:
                type: string
              storageServiceResponse:
                type: string
            example:
              message: "Unidentified error"
              error: ""
              storageServiceResponse: "Message from Storage Service"
        400:
          description: "Invalid Request, when parameters are missing"
          schema:
            properties:
              message:
                type: string
            example:
              message: "Invalid Request"
        408:
          description: "Storage Service timed out"
          schema:
            properties:
              message:
                type: string
            example:
              message: "Fetching from Storage Service timed out"
  /latest:
    get:
      tags:
      - "Search"
      summary: "Gets 50 latest memes"
      description: "Gets 50 latest memes"
      produces:
      - "application/json"
      responses:
        200:
          description: "Successful operation"
          schema:
            $ref: "#/definitions/query_response"
        500:
          description: "Service error"
          schema:
            properties:
              message:
                type: string
              error: 
                type: string
              storageServiceResponse:
                type: string
            example:
              message: "Unidentified error"
              error: ""
              storageServiceResponse: "Message from Storage Service"
        408:
          description: "Storage Service timed out"
          schema:
            properties:
              message:
                type: string
            example:
              message: "Fetching from Storage Service timed out"
...
  /meme/{memeid}:
    get:
      tags:
      - "Meme"
      summary: "Get meme by id"
      description: "Get the memes from the storage and send it as a file."
      produces:
      - "file"
      parameters:
      - in: "path"
        name: "memeid"
        required: true
        type:
          string
      responses:
        200:
          description: "Return the meme image."
          schema:
            properties:
             meme:
              type: string
            example:
              file
        404:
          description: "Meme not found"
          schema:
            properties:
              message:
                type: string
            example:
              message: "Meme not found"
        500:
          description: "Service error"
          schema:
            properties:
              message:
                type: string
              error: 
                type: string
              storageServiceResponse:
                type: string
            example:
              message: "Unidentified error"
              error: ""
              storageServiceResponse: "Message from Storage Service"
        408:
          description: "Storage Service timed out"
          schema:
            properties:
              message:
                type: string
            example:
              message: "Fetching from Storage Service timed out"
        400:
          description: "Invalid Request, when parameters are missing"
          schema:
            properties:
              message:
                type: string
            example:
              message: "Invalid Request"
...
definitions:
  query_response:
    required: 
    - "memeIds"
    - "message"
    properties:
      memeIds:
        type: array
        items:
          type: string
      message:
        type: string
    example: 
      1: 
        createdAt: 14.03.2020
        tags: ["junior", "git", "python"]
      2:
        createdAt: 19.03.2020
        tags: ["js", "dev", "ops"]
...